Revert "move init func to separate files"
This reverts commit 3de50a64129fb6da93a8792de6d55c22aa4d5373.
