+++ app/commander/commander.go
@@ -100,3 +100,9 @@ func (c *Commander) Close() error {
 
 	return nil
 }
+
+func init() {
+	common.Must(common.RegisterConfig((*Config)(nil), func(ctx context.Context, cfg interface{}) (interface{}, error) {
+		return NewCommander(ctx, cfg.(*Config))
+	}))
+}
