--- app/commander/init.go
@@ -1,15 +1,0 @@
-// +build !confonly
-
-package commander
-
-import (
-	"context"
-
-	"v2ray.com/core/common"
-)
-
-func init() {
-	common.Must(common.RegisterConfig((*Config)(nil), func(ctx context.Context, cfg interface{}) (interface{}, error) {
-		return NewCommander(ctx, cfg.(*Config))
-	}))
-}
