+++ transport/internet/quic/quic.go
@@ -1,5 +1,10 @@
 package quic
 
+import (
+	"v2ray.com/core/common"
+	"v2ray.com/core/transport/internet"
+)
+
 //go:generate errorgen
 
 // Here is some modification needs to be done before update quic vendor.
@@ -10,3 +15,9 @@ package quic
 
 const protocolName = "quic"
 const internalDomain = "quic.internal.v2ray.com"
+
+func init() {
+	common.Must(internet.RegisterProtocolConfigCreator(protocolName, func() interface{} {
+		return new(Config)
+	}))
+}
